Eliminate 'comparing signed/unsigned values' lcc warning in add_chunk_as (a cherry-pick of commit dd584df18 from 'release-7_8')
* src/atomic_ops_malloc.c (add_chunk_as): Cast CHUNK_SIZE to size_t
when compared to value of sz local variable.

+++ src/atomic_ops_malloc.c
@@ -233,7 +233,7 @@ static void add_chunk_as(void * chunk, unsigned log_sz)
   size_t ofs, limit;
   size_t sz = (size_t)1 << log_sz;
 
-  assert (CHUNK_SIZE >= sz);
+  assert((size_t)CHUNK_SIZE >= sz);
   limit = (size_t)CHUNK_SIZE - sz;
   for (ofs = ALIGNMENT - sizeof(AO_t); ofs <= limit; ofs += sz) {
     AO_stack_push(&AO_free_list[log_sz], (AO_t *)((char *)chunk + ofs));
